Replace old Virus 1 with new mesh + anim
Also reorganizes the folder content to better group things.  There's a lot of
'new file' and 'deleted' messages, but those are just file moves.  The reason
there's no 'renamed' messages is because UE4 has to tweak the binary files to
get them to point to the new location, so they're different to Git.