Finishes Virus 3 mesh & anim
All the materials and maps are created, the virus is exported, and it is
imported into UE4.

I'M REALLY SALTY that UE4 doesn't handle non-uniform scale animations for
verticies.  If it does, I couldn't figure it out.  So, RELCUTANTLY... I'm just
making the head grow and shrink as the animation...  I spent soooo long trying
to get that animation just right and it looked awesome!

Lesson learned:  make animations with bones and skin... *sigh*