Recreate level 1 and refactor room dir structure
Lots of changes.  Rooms 1-6 are removed, and Level1 is renamed to
PersistentLevel.  Going forward, each "room" will be named "Level" instead of
"Room" because it makes more sense that way.

On that note, Level 1 is blocked out with custom meshes for the walls.  Is is
completely barren, so there will be more assets created in the future.

Thirdly, we are no longer using teleporters between levels.  It will be an
elevator system, which means that streaming volumes for levels are going to be
used.